refactor(useFetchCartItems): ShoppingCart 컴포넌트의 데이터 처리 로직을 훅으로 분리
